input: misc: drv2624_haptic: Rename to qti-haptics
This driver behaves like the qti haptics driver and hence we can
use the qti vibrator HAL. Since [1] the qti vibrator HAL checks
wether a qti haptics device is detected and does not work if it's
not a qti one. In order to allow the HAL to work with this driver
we spoof the input device name to qti-haptics if the actual
qti-haptics driver is not built.

[1] https://github.com/LineageOS/android_vendor_qcom_opensource_vibrator/commit/c8f31f143e5ed2bfed4f47193fc7514580e87220

Change-Id: Ia41019a2dfcc54f0170ffe8877d19102b3b0fb83
Signed-off-by: Subhajeet Muhuri <subhajeet.muhuri@aosip.dev>